Removed utils files, spacing and other changes.
