+++ src/PartialModelSelection.cpp
@@ -7,7 +7,7 @@
 #include <string>
 //Local includes
 #include "PartialModelSelection.hpp"
-#include "ModelSelectionUtils.hpp"
+
 
 
 /*MODEL SELECTION MAP IMPLEMENTATIONS*/
@@ -28,8 +28,6 @@ ModelSelectionMap::ModelSelectionMap() : maxModels(STD_MODEL_CAP){
    catch(std::logic_error errorMessage) {
       std::cout << "Insert failed, validation returned an error!\n";
    }
-
-
    insertedModels = 0; //This starts at 0 as we exclude the beginning placeholder. 
 }
 
@@ -41,21 +39,15 @@ ModelSelectionMap::ModelSelectionMap(int maxModels) : maxModels(maxModels) {
    try{
       auto insertResult = penaltyModelMap.insert(startingPair);
       validateInsert(insertResult);
-      prevInsertedPair = insertResult.first; //Set the iterator to the previous insert to the validated insert as we did not error.
-      
+      prevInsertedPair = insertResult.first; //Set the iterator to the previous insert to the validated insert as we did not error.  
    }
-
    catch(std::logic_error errorMessage) {
       std::cout << "Insert failed, validation returned an error!\n";
    }
-   
-   
-   
    insertedModels = 0; //This starts at 0 as we exclude the beginning placeholder.
 }
 
-
-void ModelSelectionMap::insert(double newPenalty, Model newModel){
+ void ModelSelectionMap::insert(double newPenalty, Model newModel){
    //Insert into ourpenaltyModelPair map in the ModelSelectionMap if the newPenalty is not within it.
    PenaltyModelPair newPair = PenaltyModelPair(newPenalty, newModel);
    
@@ -70,22 +62,8 @@ void ModelSelectionMap::insert(double newPenalty, Model newModel){
    catch(std::logic_error errorMessage) {
       std::cout << "Insert failed, validation returned an error!\n";
    }
-
-   
-   
    insertedModels++;     
 }
-
-
-void ModelSelectionMap::insert(Model currentModel)
-{
-   //Runs solver for penaltyQuery to insert
-   //std::cout << "Adding model without penaltyQuery \n";
-
-   //Insert pair with createdpenaltyModelPair from solver.
-}
-
-
 
 MinimizeResult ModelSelectionMap::minimize(double penaltyQuery){
    //Default result if we do not find a matching penaltyQuery.
@@ -99,14 +77,14 @@ MinimizeResult ModelSelectionMap::minimize(double penaltyQuery){
     std::cout << "Map is empty, returning default min result\n";
     return queryResult;
    }
-
     //If we have models inserted, we need to find a penaltyQuery result that is closest to the queried penaltyQuery
     indexPair = penaltyModelMap.lower_bound(penaltyQuery);
 
     //If we found an inserted pair that lies on the queried penalty itself
     if(indexPenalty == penaltyQuery)
-      //Make a query result to return using the second element of a testedPair, Model. Get it's modelSize. 
-      queryResult = MinimizeResult(std::make_pair(indexModel.modelSize,indexModel.modelSize));
+    
+    //Make a query result to return using the second element of a testedPair, Model. Get it's modelSize. 
+    queryResult = MinimizeResult(std::make_pair(indexModel.modelSize,indexModel.modelSize));
    
     //Otherwise, make a range query that does not lie on an inserted pair. 
     else{
@@ -117,39 +95,20 @@ MinimizeResult ModelSelectionMap::minimize(double penaltyQuery){
     return queryResult;
 }
 
-
-
-
-double ModelSelectionMap::getNewPenaltyList()
-{
-
+double ModelSelectionMap::getNewPenalty(){
   //If the map of tested pairs is empty, query penaltyQuery 0 first.  
-  if(!hasModelsInserted())
-     {
+  if(!hasModelsInserted()){
        return EMPTY_MAP_QUERY; //?
-     }
-
-  //penaltyModelpenaltyModelPair smallestPair = penaltyModelMap.begin()->first;
-  
-  //for (std::map<double,Model>::iterator it=penaltyModelMap.begin(); it!=penaltyModelMap.end(); ++it)
-    //  std::cout << it->first << " => " << it->second.modelSize << '\n';   
-   return 0;
-     
-    
-  
+  }
+   return 0; 
 }
 
-
-
-std::pair<int, int> ModelSelectionMap::solver(double penaltyQuery)
-{
+std::pair<int, int> ModelSelectionMap::solver(double penaltyQuery){
    auto tempPair = std::make_pair<int, bool>(4, true);
    std::cout << "Testing solver!\n";
    //Temporary stub return
    return tempPair;
 }
-
-
 
 //Method to display the currently stored pairs in the map. 
 void ModelSelectionMap::displayMap() {
@@ -161,8 +120,6 @@ void ModelSelectionMap::displayMap() {
    std::cout << " \n";
  }
   
-
-
  bool ModelSelectionMap::hasModelsInserted(){
     //Check to see if only the default entry is in the map.
     return insertedModels > 0;
@@ -174,8 +131,7 @@ void ModelSelectionMap::displayMap() {
 }
 
 
- //General Utilities used in ModelSelectionMap
-
+//General Utilities used in ModelSelectionMap
 //Takes in an insertion result and returns the iterator to the insertion if it is valid. 
 std::map<double, Model>::iterator validateInsert(std::pair<std::map<double, Model>::iterator, bool> insertResult){
     //The result from insert is: std::pair<iterator,bool> where the bool represents the success/failure of insertion.
@@ -184,6 +140,41 @@ std::map<double, Model>::iterator validateInsert(std::pair<std::map<double, Mode
     return validIterator;
 }
 
+
+//MinimizeResult Method Implementations
+//MinimizeResult default constructor to be used when no models have been inserted. 
+MinimizeResult::MinimizeResult(){
+ //Initialize a default result, which has 1 selected, but unsure.
+ optimalModels.first = 1;
+ optimalModels.second = INFINITY;
+ certain = false;
+
+}
+
+MinimizeResult::MinimizeResult(std::pair<double, double> inputModels){
+       //Check for a valid range of models. 
+       if(!isValidRange(inputModels)) {
+           throw std::out_of_range("Invalid range inputted to MinimizeResult creation."); 
+      }
+       else {
+           //Check if our penaltyQueryRange is a solved point (Where the beginning and end are identical.)
+           if(optimalModels.first == optimalModels.second)
+               certain = true;
+            else{
+               certain = false;  
+            }
+            optimalModels.first = inputModels.first;
+            optimalModels.second = inputModels.second;
+      }
+   }
+
+bool MinimizeResult::isValidRange(std::pair<double,double> optimalModels){
+    //Tie a model size and pairing to a penaltyQuery range.
+    if(optimalModels.first < 0 || optimalModels.first < optimalModels.second){
+        return false;   
+    }
+    return true;
+}
